requirements.txt
pandas
streamlit
python-dotenv
langchain==0.2.5
langchain-community==0.2.5
langchain-google-genai==1.0.6